% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nba_data.R
\name{nba_data_base}
\alias{nba_data_base}
\title{NBA Data Base}
\format{
A \code{tibble} with the following columns:
\describe{
  \item{\code{name}}{The name of the player.}
  \item{\code{year}}{The year of the draft.}
  \item{\code{round}}{The round in which the player was drafted.}
}
}
\usage{
nba_data_base()
}
\value{
A filtered and selected tibble of NBA data from Base.
}
\description{
Filter NBA Data by Source (Base)
}
\details{
Filters and selects NBA data from the Base (ESPN.com) source.
}
\examples{
# Filter NBA data for base source
base_data <- nba_data_base()

# View the first few rows
head(base_data)

}
