% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nba_data.R
\name{nba_data_espn}
\alias{nba_data_espn}
\title{NBA Data ESPN}
\usage{
nba_data_espn()
}
\value{
A filtered and selected tibble of NBA data from ESPN.
}
\description{
Filter NBA Data by Source (ESPN)
}
\details{
Filters and selects NBA data from ESPN for the given source value.
}
