% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nba_data.R
\name{nba_data_nba_com}
\alias{nba_data_nba_com}
\title{NBA Data NBA.com}
\usage{
nba_data_nba_com()
}
\value{
A filtered and selected tibble of NBA data from NBA.com.
}
\description{
Filter NBA Data by Source (NBA.com)
}
\details{
Filters and selects NBA data from NBA.com for the given source value.
}
