% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nba_data.R
\name{nba_data_nbadraft_net}
\alias{nba_data_nbadraft_net}
\title{NBA Data NBADraft.net}
\usage{
nba_data_nbadraft_net()
}
\value{
A filtered and selected tibble of NBA data from NBADraft.net.
}
\description{
Filter NBA Data by Source (NBADraft.net)
}
\details{
Filters and selects NBA data from NBADraft.net for the given source value.
}
