% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nba_data.R
\name{nfl_data_nba_com}
\alias{nfl_data_nba_com}
\title{NFL Data NBA.com}
\usage{
nfl_data_nba_com()
}
\value{
A filtered and selected tibble of NFL data from NBA.com.
}
\description{
Filter NFL Data by Source (NBA.com)
}
\details{
Filters and selects NFL data from NBA.com for the given source value.
}
