% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nba_data.R
\name{nfl_data_nbadraft_net}
\alias{nfl_data_nbadraft_net}
\title{NFL Data NBADraft.net}
\usage{
nfl_data_nbadraft_net()
}
\value{
A filtered and selected tibble of NFL data from NBADraft.net.
}
\description{
Filter NFL Data by Source (NBADraft.net)
}
\details{
Filters and selects NFL data from NBADraft.net for the given source value.
}
