% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nba_data.R, R/nfl_data.R
\name{nfl_data_the_ringer}
\alias{nfl_data_the_ringer}
\title{NFL Data The Ringer}
\usage{
nfl_data_the_ringer()

nfl_data_the_ringer()
}
\value{
A filtered and selected tibble of NFL data from The Ringer.

A filtered and selected tibble of NFL data from The Ringer.
}
\description{
Filter NFL Data by Source (The Ringer)

Filter NFL Data by Source (The Ringer)
}
\details{
Filters and selects NFL data from The Ringer for the given source value.

Filters and selects NFL data from The Ringer for the given source value.
}
