% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nfl_data.R
\name{nfl_data_walter_football}
\alias{nfl_data_walter_football}
\title{NFL Data Walter Football}
\usage{
nfl_data_walter_football()
}
\value{
A filtered and selected tibble of NFL data from Walter Football.
}
\description{
Filter NFL Data by Source (Walter Football)
}
\details{
Filters and selects NFL data from Walter Football for the given source value.
}
