% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhl_data.R
\name{nhl_data_base}
\alias{nhl_data_base}
\title{NHL Data Base}
\usage{
nhl_data_base(source_value = "Base")
}
\arguments{
\item{source_value}{A character string specifying the source of the data. Default is "Base".}
}
\value{
A filtered and selected tibble of NHL data.
}
\description{
Filter NHL Data by Source (Base)
}
\details{
Filters and selects NHL data from the base source for the given source value.
}
