% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhl_data.R
\name{nhl_data_espn}
\alias{nhl_data_espn}
\title{NHL Data ESPN}
\usage{
nhl_data_espn(source_value = "ESPN")
}
\arguments{
\item{source_value}{A character string specifying the source of the data. Default is "ESPN".}
}
\value{
A filtered and selected tibble of NHL data from ESPN.
}
\description{
Filter NHL Data by Source (ESPN)
}
\details{
Filters and selects NHL data from ESPN for the given source value.
}
