% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{output}
\alias{output}
\title{Silly Printer function}
\usage{
output(x, y)
}
\arguments{
\item{x}{what you want in second column}

\item{y}{what you want in first  column}
}
\value{
A tibble
}
\description{
Silly Printer function
}
\examples{
output(x = rnorm(5), y = rnorm(5))
}
