% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProSportsDraftData-package.R
\docType{package}
\name{ProSportsDraftData-package}
\alias{ProSportsDraftData}
\alias{ProSportsDraftData-package}
\title{ProSportsDraftData: Professional Sports Draft Data}
\description{
The `ProSportsDraftData` package provides comprehensive draft data for major professional sports leagues, including the National Football League (NFL), National Basketball Association (NBA), and National Hockey League (NHL). It offers access to both historical and current draft data, allowing for detailed analysis and research on player biases and player performance. The package is useful for sports fans and researchers interested in identifying biases and trends within scouting reports. Created by web scraping data from leading websites that cover professional sports player scouting reports, the package allows users to filter and summarize data for analytical purposes. For further details on the methods used, please refer to Wickham (2022) "rvest: Easily Harvest (Scrape) Web Pages" https://CRAN.R-project.org/package=rvest and Harrison (2023) "RSelenium: R Bindings for Selenium WebDriver" \url{https://CRAN.R-project.org/package=RSelenium}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Ginsburg1/ProSportsDraftData}
  \item Report bugs at \url{https://github.com/Ginsburg1/ProSportsDraftData/issues}
}

}
\author{
\strong{Maintainer}: Benjamin Ginsburg \email{benjamin.ginsburg@du.edu}

}
\keyword{internal}
