% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nba_data.R
\name{nba_data_base}
\alias{nba_data_base}
\title{NBA Data Base}
\usage{
nba_data_base()
}
\value{
A filtered and selected tibble of NBA data from Base.
}
\description{
Filter NBA Data by Source (Base)
}
\details{
Filters and selects NBA data from the Base (ESPN.com) source.
}
