% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nba_data.R
\name{nba_data_the_ringer}
\alias{nba_data_the_ringer}
\title{NBA Data The Ringer}
\usage{
nba_data_the_ringer()
}
\value{
A filtered and selected tibble of NBA data from The Ringer.
}
\description{
Filter NBA Data by Source (The Ringer)
}
\details{
Filters and selects NBA data from The Ringer for the given source value.
}
