% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nfl_data.R
\name{nfl_data_base}
\alias{nfl_data_base}
\title{NFL Data Base}
\format{
A \code{tibble} with the following columns:
\describe{
  \item{\code{name}}{The name of the player.}
  \item{\code{round}}{The round in which the player was drafted.}
  \item{\code{rank}}{The rank of the player.}
}
}
\usage{
nfl_data_base()
}
\value{
A filtered and selected tibble of NFL data.
}
\description{
Filter NFL Data by Source (Base)
}
\details{
Filters and selects NFL data from the base source for the given source value.
}
\examples{
# Filter NFL data for base source
base_data <- nfl_data_base()

# View the first few rows
head(base_data)

}
