% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nba_data.R, R/nfl_data.R
\name{nfl_data_base}
\alias{nfl_data_base}
\title{NFL Data Base}
\usage{
nfl_data_base()

nfl_data_base()
}
\value{
A filtered and selected tibble of NFL data from Base.

A filtered and selected tibble of NFL data.
}
\description{
Filter NFL Data by Source (Base)

Filter NFL Data by Source (Base)
}
\details{
Filters and selects NFL data from the Base (ESPN.com) source.

Filters and selects NFL data from the base source for the given source value.
}
