% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nba_data.R, R/nfl_data.R
\name{nfl_data_espn}
\alias{nfl_data_espn}
\title{NFL Data ESPN}
\usage{
nfl_data_espn()

nfl_data_espn()
}
\value{
A filtered and selected tibble of NFL data from ESPN.

A filtered and selected tibble of NFL data from ESPN.
}
\description{
Filter NFL Data by Source (ESPN)

Filter NFL Data by Source (ESPN)
}
\details{
Filters and selects NFL data from ESPN for the given source value.

Filters and selects NFL data from ESPN for the given source value.
}
