% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nfl_data.R
\name{nfl_data_nfl_com}
\alias{nfl_data_nfl_com}
\title{NFL Data NFL.com}
\usage{
nfl_data_nfl_com()
}
\value{
A filtered and selected tibble of NFL data from NFL.com.
}
\description{
Filter NFL Data by Source (NFL.com)
}
\details{
Filters and selects NFL data from NFL.com for the given source value.
}
