% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhl_data.R
\name{nhl_data_base}
\alias{nhl_data_base}
\title{NHL Data Base}
\format{
A \code{tibble} with the following columns:
\describe{
  \item{\code{name}}{The name of the player.}
  \item{\code{round}}{The round in which the player was drafted.}
  \item{\code{rank}}{The rank of the player in the draft.}
}
}
\usage{
nhl_data_base()
}
\value{
A filtered and selected tibble of NHL data.
}
\description{
Filter NHL Data by Source (Base)
}
\details{
Filters and selects NHL data from the base source for the given source value.
}
\examples{
# Filter NHL data for base source
base_data <- nhl_data_base()

# View the first few rows
head(base_data)

}
