% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhl_data.R
\name{nhl_data_nhl_com}
\alias{nhl_data_nhl_com}
\title{NHL Data NHL.com}
\usage{
nhl_data_nhl_com()
}
\value{
A filtered and selected tibble of NHL data from NHL.com.
}
\description{
Filter NHL Data by Source (NHL.com)
}
\details{
Filters and selects NHL data from NHL.com for the given source value.
}
