% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProSportsDraftData-package.R
\docType{package}
\name{ProSportsDraftData-package}
\alias{ProSportsDraftData}
\alias{ProSportsDraftData-package}
\title{ProSportsDraftData: Professional Sports Draft Data}
\description{
Provides draft data from various sources for professional sports leagues, including the NFL, NBA, and NHL.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Ginsburg1/ProSportsDraftData}
  \item Report bugs at \url{https://github.com/Ginsburg1/ProSportsDraftData/issues}
}

}
\author{
\strong{Maintainer}: Benjamin Ginsburg \email{benjamin.ginsburg@du.edu}

}
\keyword{internal}
